+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/docker/dataprovider/ContainerDAO.java
@@ -7,9 +7,5 @@ import java.util.List;
 
 public interface ContainerDAO {
 
-  boolean isDpsContainer(String containerId);
-
-  Container findDPSContainer();
-
   List<Container> findActiveContainersInspectMatching(HostnameQuery query);
 }
