+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/docker/dataprovider/DpsContainerDAO.java
@@ -1,0 +1,15 @@
+package com.mageddo.dnsproxyserver.server.dns.solver.docker.dataprovider;
+
+import com.mageddo.dnsproxyserver.server.dns.solver.docker.Container;
+
+public interface DpsContainerDAO {
+
+
+  boolean isDpsContainer(String containerId);
+
+  Container findDPSContainer();
+
+  boolean isDpsRunningInsideContainer();
+
+  void createDpsNetwork();
+}
